refactor: move reservations page to spa/
