+++ apps/admin-ui/src/component/Legend.tsx
