+++ apps/admin-ui/src/hooks/useCheckCollisions.ts
@@ -9,7 +9,7 @@ import { base64encode, filterNonNullable } from "common/src/helpers";
 import { toApiDate } from "common/src/common/util";
 import { RELATED_RESERVATION_STATES } from "common/src/const";
 
-function useCheckCollisions({
+export function useCheckCollisions({
   reservationPk,
   reservationUnitPk,
   start,
@@ -86,5 +86,3 @@ function useCheckCollisions({
 
   return { isLoading: loading, hasCollisions: collisions.length > 0 };
 }
-
-export default useCheckCollisions;
