+++ apps/admin-ui/src/spa/my-units/[id]/hooks/queries.tsx
@@ -2,9 +2,9 @@ import { gql } from "@apollo/client";
 import {
   RESERVATIONUNIT_RESERVATIONS_FRAGMENT,
   RESERVATION_UNIT_FRAGMENT,
-} from "@/component/reservations/fragments";
+  UNIT_NAME_FRAGMENT,
+} from "@/common/fragments";
 import { LOCATION_FRAGMENT } from "common/src/queries/fragments";
-import { UNIT_NAME_FRAGMENT } from "@/common/fragments";
 
 export const OPTIONS_QUERY = gql`
   query Options {
