+++ apps/admin-ui/src/spa/reservations/requested/ApprovalButtonsRecurring.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { ReservationStateChoice, type ReservationQuery } from "@gql/gql-types";
-import { useTranslation } from "react-i18next";
+import { useTranslation } from "next-i18next";
 import { Button } from "hds-react";
 import { ButtonLikeLink } from "@/component/ButtonLikeLink";
-import DenyDialog from "./DenyDialog";
+import DenyDialog from "@/component/DenyDialog";
 import { useModal } from "@/context/ModalContext";
 import { useRecurringReservations } from "./hooks";
