+++ apps/admin-ui/src/spa/reservations/requested/Calendar.tsx
@@ -6,8 +6,8 @@ import { useTranslation } from "react-i18next";
 import { type ReservationQuery, ReservationTypeChoice } from "@gql/gql-types";
 import { useModal } from "@/context/ModalContext";
 import eventStyleGetter, { legend } from "./eventStyleGetter";
-import Legend from "./Legend";
-import { EditTimeModal } from "../EditTimeModal";
+import Legend from "@/component/Legend";
+import { EditTimeModal } from "@/component/EditTimeModal";
 import { isPossibleToEdit } from "./reservationModificationRules";
 import { getEventBuffers } from "common/src/calendar/util";
 import { filterNonNullable } from "common/src/helpers";
