+++ apps/admin-ui/src/spa/reservations/requested/eventStyleGetter.ts
