+++ apps/admin-ui/src/spa/reservations/requested/hooks/index.ts
@@ -3,19 +3,16 @@ import {
   ReservationStateChoice,
   ReservationTypeChoice,
   useReservationQuery,
-  useReservationDenyReasonsQuery,
   useReservationsByReservationUnitQuery,
   useRecurringReservationQuery,
   ReservationUnitNode,
   ReservationNode,
 } from "@gql/gql-types";
-import { useTranslation } from "react-i18next";
+import { useTranslation } from "next-i18next";
 import { toApiDate } from "common/src/common/util";
 import { errorToast } from "common/src/common/toast";
 import { base64encode, filterNonNullable } from "common/src/helpers";
 import { type CalendarEventType } from "../Calendar";
-
-export { default as useCheckCollisions } from "./useCheckCollisions";
 
 const getEventName = (
   eventType?: ReservationTypeChoice | null,
@@ -158,28 +155,6 @@ export function useRecurringReservations(recurringPk?: number) {
     recurringReservation,
     refetch,
   };
-}
-
-// TODO this has the same useState being local problems as useRecurringReservations
-// used to have but it's not obvious because we don't mutate / refetch this.
-// Cache it in Apollo InMemory cache instead.
-export function useDenyReasonOptions() {
-  const { t } = useTranslation();
-
-  const { data, loading } = useReservationDenyReasonsQuery({
-    onError: () => {
-      errorToast({ text: t("errors.errorFetchingData") });
-    },
-  });
-  const { reservationDenyReasons } = data ?? {};
-  const denyReasonOptions = filterNonNullable(
-    reservationDenyReasons?.edges.map((x) => x?.node)
-  ).map((dr) => ({
-    value: dr?.pk ?? 0,
-    label: dr?.reasonFi ?? "",
-  }));
-
-  return { options: denyReasonOptions, loading };
 }
 
 /// @param id fetch reservation related to this pk
