+++ apps/admin-ui/src/spa/reservations/requested/hooks/queries.tsx
@@ -1,10 +1,12 @@
 import { gql } from "@apollo/client";
 import {
-  RESERVATION_COMMON_FRAGMENT,
   RESERVATION_RECURRING_FRAGMENT,
-  RESERVATION_UNIT_FRAGMENT,
   RESERVATION_UNIT_PRICING_FRAGMENT,
 } from "../../fragments";
+import {
+  RESERVATION_COMMON_FRAGMENT,
+  RESERVATION_UNIT_FRAGMENT,
+} from "@/common/fragments";
 import {
   RESERVEE_BILLING_FRAGMENT,
   RESERVEE_NAME_FRAGMENT,
