+++ apps/admin-ui/src/spa/reservations/requested/util.ts
@@ -112,14 +112,6 @@ export function getReservationPriceDetails(
       });
 }
 
-export const ageGroup = (
-  group:
-    | Maybe<
-        Pick<NonNullable<ReservationType["ageGroup"]>, "minimum" | "maximum">
-      >
-    | undefined
-): string | null => (group ? `${group.minimum}-${group.maximum || ""}` : null);
-
 const reserveeTypeToTranslationKey = (
   reserveeType: CustomerTypeChoice,
   isUnregisteredAssociation: boolean
